changed the increments from normal to big one

+++ src/Database/Migrations/2018_05_14_184308_create_dz_page_definitions_table.php
@@ -16,15 +16,15 @@ class CreateDzPageDefinitionsTable extends Migration {
         Schema::create('dz_page_definitions', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('name');
             $table->string('path')->nullable();
 
             $table->string('description')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
