+++ src/Database/Migrations/2018_06_01_123440_create_dz_module_details_table.php
@@ -16,15 +16,15 @@ class CreateDzModuleDetailsTable extends Migration {
         Schema::create('dz_module_details', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('name');
             $table->string('image')->nullable();
             $table->string('description')->nullable();
             $table->integer('display_order');
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
