+++ src/Database/Migrations/2018_06_21_124258_create_dz_ui_actions_table.php
@@ -16,7 +16,7 @@ class CreateDzUiActionsTable extends Migration {
         Schema::create('dz_ui_actions', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->bigIncrements('id');
+            $table->increments('id');
 
             $table->string('name');
             $table->string('identifier')->nullable();
@@ -24,7 +24,7 @@ class CreateDzUiActionsTable extends Migration {
             $table->string('description')->nullable();
 
             $table->string('source_type')->nullable();
-            $table->unsignedBigInteger('source_id')->nullable();
+            $table->unsignedInteger('source_id')->nullable();
 
             $table->integer('display_order')->default(0);
             $table->string('image')->nullable();
@@ -37,12 +37,12 @@ class CreateDzUiActionsTable extends Migration {
 
             $table->boolean('active')->default(true);
 
-            $table->unsignedBigInteger('form_id')->nullable();
-            $table->unsignedBigInteger('filter_condition_id')->nullable();
-            $table->unsignedBigInteger('execution_script_id')->nullable();
+            $table->unsignedInteger('form_id')->nullable();
+            $table->unsignedInteger('filter_condition_id')->nullable();
+            $table->unsignedInteger('execution_script_id')->nullable();
 
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->foreign('form_id')->references('id')->on('dz_custom_forms');
             $table->foreign('filter_condition_id')->references('id')->on('dz_system_scripts');
